pyproject.toml
[project]
name = "pdf2docx-md"
version = "0.1.0"
description = "High-fidelity PDF to Markdown and DOCX conversion toolkit with OCR support."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdf2docx>=0.5.8",
    "pdf2image>=1.17.0",
    "pillow>=12.1.0",
    "pypdf>=6.6.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
]
